+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Implement backward training.
 - Update the method `myNetwork::initWeight()` to reduce the initialized weights.
 - Implement momentum SGD.
+- Implement the method `myNetwork::calcRecall()` to calculate recall.
 
 ### Changed